Replace NetworkingAnalytics with Analytics
